chore(scripts): archive 6 one-off scripts to reduce count below 35 (#2343)
Move backfill_judge_department, backfill_normalize_departments,
cleanup_orphaned_documents, dedup_judges, migrate_directory_s3_keys,
and remediate_enrichment_fixtures to scripts/archive/. All are marked
with `# one-off: true` and their work is complete. Updates doc examples
to reference still-active scripts.

Script count: 38 -> 32 (threshold: 35).
