+++ scripts/archive/cleanup_orphaned_documents.py
