+++ scripts/archive/migrate_directory_s3_keys.py
